Remove yiisoft/strings parts 2
